Refactor: Move store directory for consistency
