+++ src/app/(Main)/EDT/ui/schedule-form.tsx
@@ -24,8 +24,8 @@ import {getRooms} from "@/services/Room";
 import {Group} from "@/Types/Group";
 import {getAvailableGroups} from "@/Tools/Group";
 import {generateHours} from "@/Tools/ScheduleItem";
-import {useCurrentScheduleItemsStore} from "@/app/Stores/ScheduleItem";
-import {useCurrentLevelStore} from "@/app/Stores/Level";
+import {useCurrentScheduleItemsStore} from "@/Stores/ScheduleItem";
+import {useCurrentLevelStore} from "@/Stores/Level";
 import {Dialog, DialogContent, DialogHeader, DialogTitle} from "@/components/ui/dialog";
 import {ScrollArea} from "@/components/ui/scroll-area";
 import {addScheduleItemService} from "@/services/ScheduleItem";
